+++ CHANGELOG.md
@@ -24,7 +24,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Default failure mode changed from `continue` to `no-new` (see above for
   definitions).
 
-- A distinct event is now logged when a script is terminated intentionally by
+- A distinct event is now logged when a script is killed intentionally by
   Wireit.
 
 ## [0.4.1] - 2022-05-10